// src/app/models/breedDto.ts
export class BreedDto {
    id: number = 0;

    mainPhoto: string = "";
    mainDescription: string = "";

    bigPhoto: string = "";
    bigDescription: string = "";

    breed: string = "";
    country: string = "";
    weight: string = "";
    height: string = "";
    lifeExpectancy: string = "";

    aggressiveness: number = 0;
    activity: number = 0;
    health: number = 0;
    intelligence: number = 0;
    training: number = 0;
    maintenanceCost: number = 0;
    molting: number = 0;
    securityQualities: number = 0;
    price: string = '';
    noise: number = 0;
    
    dogType: string = "";
}